pyproject.toml
[tool.poetry]
name = "underkingbot"
version = "0.1.0"
description = ""
authors = ["SilentWrangler <thousandgoblins@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.4"
discord-py-interactions = "^5.9.2"
python-dotenv = "^1.0.0"
dice = "^4.0.0"
Pillow = "^10.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
